requirements.txt
aiohttp==3.9.3
SpeechRecognition==3.10.1
pyttsx3==2.90
python-dotenv==1.0.1
PyAudio==0.2.14 